refactor: 하나로 합쳐져 있던 테스트 함수를 각 클래스별 단위테스트로 분리
